Move PredictionMode and related models to predict.rs
