+++ src/context.rs
@@ -18,9 +18,10 @@ use crate::encoder::FrameInvariants;
 use crate::entropymode::*;
 use crate::header::ReferenceMode;
 use crate::partition::BlockSize::*;
-use crate::partition::PredictionMode::*;
 use crate::partition::RefType::*;
 use crate::partition::*;
+use crate::predict::PredictionMode;
+use crate::predict::PredictionMode::*;
 use crate::transform::TxSize::*;
 use crate::transform::TxType::*;
 use crate::transform::*;
