+++ src/deblock.rs
@@ -14,9 +14,9 @@ use crate::DeblockState;
 use crate::encoder::FrameInvariants;
 use crate::encoder::FrameState;
 use crate::api::FrameType;
-use crate::partition::PredictionMode::*;
 use crate::partition::RefType::*;
 use crate::plane::*;
+use crate::predict::PredictionMode::*;
 use crate::quantize::*;
 use crate::util::Pixel;
 use std::cmp;
